+++ Scripts/New-FGBranchExplorer.ps1
@@ -3,6 +3,8 @@ function New-FGBranchExplorer($Root) {
 	New-Object PowerShellFar.PowerExplorer 71ebb500-c8e1-4544-827a-8456c3611f8e -Property @{
 		Data = @{
 			Root = $Root
+			SetCurrentOnce = $true
+			FileDescriptions = @{}
 		}
 		Functions = 'CreateFile, DeleteFiles, RenameFile'
 		AsCreateFile = {FGBranchExplorer_AsCreateFile @args}
@@ -13,34 +15,42 @@ function New-FGBranchExplorer($Root) {
 	}
 }
 
-function FGBranchExplorer_AsCreatePanel($1) {
-	$panel = [FarNet.Panel]$1
-	$panel.Title = "Branches: $($1.Data.Root)"
+function FGBranchExplorer_AsCreatePanel($Explorer) {
+	$panel = [FarNet.Panel]$Explorer
+	$panel.Title = "Branches: $($Explorer.Data.Root)"
 	$panel.ViewMode = 0
 	$panel.SortMode = 'Unsorted'
+	$Explorer.Data.Panel = $panel
 
-	$co = New-Object FarNet.SetColumn -Property @{ Kind = "O"; Name = "Current"; Width = 1 }
-	$cn = New-Object FarNet.SetColumn -Property @{ Kind = "N"; Name = "Branch" }
-	$cd = New-Object FarNet.SetColumn -Property @{ Kind = "Z"; Name = "Commit" }
-	$mode = New-Object FarNet.PanelPlan
-	$mode.Columns = $co, $cn, $cd
-	$panel.SetPlan(0, $mode)
+	$co = [FarNet.SetColumn]@{ Kind = "O"; Name = "Current"; Width = 1 }
+	$cn = [FarNet.SetColumn]@{ Kind = "N"; Name = "Branch" }
+	$cd = [FarNet.SetColumn]@{ Kind = "Z"; Name = "Commit" }
+
+	$plan0 = [FarNet.PanelPlan]::new()
+	$plan0.Columns = $co, $cn, $cd
+	$panel.SetPlan(0, $plan0)
+
+	$plan1 = $plan0.Clone()
+	$plan1.IsFullScreen = $true
+	$panel.SetPlan(9, $plan1)
 
 	$panel.add_KeyPressed({
 		### [Enter] checkout branch
 		if ($_.Key.Is([FarNet.KeyCode]::Enter)) {
+			$_.Ignore = $true
 			$file = $this.CurrentFile
-			if (!$file -or $file.Owner -eq '*') {
+
+			# skip the current branch
+			if (!$file -or $file.Owner) {
 				return
 			}
-			$_.Ignore = $true
-			$Root = $this.Explorer.Data.Root
+
 			$name = $file.Name
 			if ($name.StartsWith('remotes/')) {
-				$res = Invoke-Error {git -C $Root checkout -t $name}
+				$res = Invoke-Error {git -C $this.Explorer.Data.Root checkout -t $name}
 			}
 			else {
-				$res = Invoke-Error {git -C $Root checkout $name}
+				$res = Invoke-Error {git -C $this.Explorer.Data.Root checkout $name}
 			}
 			if ($LASTEXITCODE) {
 				throw $res
@@ -51,11 +61,14 @@ function FGBranchExplorer_AsCreatePanel($1) {
 		}
 		### [F3] gitk branch
 		if ($_.Key.Is([FarNet.KeyCode]::F3)) {
+			$_.Ignore = $true
 			$file = $this.CurrentFile
-			if (!$file) {
+
+			# skip the detached branch
+			if (!$file -or $file.Name -like '(*)') {
 				return
 			}
-			$_.Ignore = $true
+
 			Push-Location -LiteralPath $this.Explorer.Data.Root
 			gitk $file.Name
 			Pop-Location
@@ -66,50 +79,88 @@ function FGBranchExplorer_AsCreatePanel($1) {
 	$panel
 }
 
-function FGBranchExplorer_AsGetFiles($1) {
-	$Root = $1.Data.Root
-	$res = Invoke-Error {git -C $Root branch -a --list --quiet}
+function FGBranchExplorer_AsGetFiles($Explorer) {
+	# use and drop files prepared by the task
+	if ($Explorer.Data.Files) {
+		$Explorer.Data.Files
+		$Explorer.Data.Files = $null
+		return
+	}
+
+	# get git branches
+	$res = Invoke-Error {git -C $Explorer.Data.Root branch -a --list --quiet}
 	if ($LASTEXITCODE) {
 		throw $res
 	}
+
+	# make files, use cached descriptions to avoid blanks in some cases
+	$Files = [System.Collections.Generic.List[object]]@()
 	foreach($line in $res) {
 		if ($line.Contains('->')) {
 			continue
 		}
 		if ($line -notmatch '^(\*)?\s*(.+)') {
-			New-FarFile -Name "Unexpected branch: '$line'"
 			continue
 		}
 		$name = $Matches[2]
-		$description = if ($name -notlike 'remotes/*') { Invoke-Error {git -C $Root log --pretty=format:%s -1 $name} }
+		$file = New-FarFile -Name $name -Description $Explorer.Data.FileDescriptions[$name]
 		if ($Matches[1]) {
-			New-FarFile -Name $name -Owner * -Description $description
+			# mark current branch
+			if (Invoke-Error {git status -s}) {
+				$file.Owner = '>'
+			}
+			else {
+				$file.Owner = '*'
+			}
+
+			# set current panel file
+			if ($Explorer.Data.SetCurrentOnce) {
+				$Explorer.Data.Panel.PostName($name)
+				$Explorer.Data.SetCurrentOnce = $false
+			}
 		}
-		else {
-			New-FarFile -Name $name -Description $description
+		$Files.Add($file)
+	}
+	$Files
+
+	# get descriptions, skip remotes and detached branches
+	Start-FarTask -Data Explorer, Files {
+		. $PSScriptRoot\Basic.ps1
+		foreach($file in $Data.Files) {
+			if ($file.Name -notlike 'remotes/*' -and $file.Name -notlike '(*)') {
+				$description = Invoke-Error {git -C $Data.Explorer.Data.Root log --pretty=format:%s -1 $file.Name}
+				$file.Description = $description
+				$Data.Explorer.Data.FileDescriptions[$file.Name] = $description
+			}
+		}
+		job {
+			$Data.Explorer.Data.Files = $Data.Files
+			$Data.Explorer.Data.Panel.Update($true)
+			$Data.Explorer.Data.Panel.Redraw()
 		}
 	}
 }
 
-function FGBranchExplorer_AsCreateFile($1, $2) {
-	$Root = $1.Data.Root
-	$oldBranch = Invoke-Error {git -C $Root branch --show-current}
+function FGBranchExplorer_AsCreateFile($Explorer, $2) {
+	$oldBranch = Invoke-Error {git -C $Explorer.Data.Root branch --show-current}
 	if ($LASTEXITCODE) {
 		throw $oldBranch
 	}
+
 	$newBranch = $Far.Input('New branch name', 'GitBranch', "New branch from $oldBranch")
 	if (!$newBranch) {
 		$2.Result = 'Ignore'
 		return
 	}
-	$res = Invoke-Error {git -C $Root checkout -b $newBranch}
+
+	$res = Invoke-Error {git -C $Explorer.Data.Root checkout -b $newBranch}
 	if ($LASTEXITCODE) {
 		$2.Result = 'Ignore'
 		Show-FarMessage "Error on create branch $newBranch`n$res" -Caption FarGit -LeftAligned -IsWarning
 	}
 }
 
-function FGBranchExplorer_AsDeleteFiles($1, $2) {
+function FGBranchExplorer_AsDeleteFiles($Explorer, $2) {
 	# ask
 	if ($2.UI) {
 		$text = @"
@@ -123,7 +174,6 @@ $($2.Files -join "`n")
 	}
 
 	# delete
-	$Root = $1.Data.Root
 	foreach($file in $2.Files) {
 		$branch = $file.Name
 		$remote = ''
@@ -133,13 +183,13 @@ $($2.Files -join "`n")
 		}
 		$res = Invoke-Error {
 			if ($remote) {
-				git -C $Root push $remote --delete $branch
+				git -C $Explorer.Data.Root push $remote --delete $branch
 			}
 			elseif ($2.Force) {
-				git -C $Root branch -D $branch
+				git -C $Explorer.Data.Root branch -D $branch
 			}
 			else {
-				git -C $Root branch -d $branch
+				git -C $Explorer.Data.Root branch -d $branch
 			}
 		}
 		if ($LASTEXITCODE) {
@@ -152,15 +202,14 @@ $($2.Files -join "`n")
 	}
 }
 
-function FGBranchExplorer_AsRenameFile($1, $2) {
-	$Root = $1.Data.Root
+function FGBranchExplorer_AsRenameFile($Explorer, $2) {
 	$oldName = $2.File.Name
 	$newName = ([string]$Far.Input('New branch name', 'GitBranch', 'Rename branch', $oldName)).Trim()
 	if (!$newName) {
 		$2.Result = 'Ignore'
 		return
 	}
-	$res = Invoke-Error {git -C $Root branch -m $oldName $newName}
+	$res = Invoke-Error {git -C $Explorer.Data.Root branch -m $oldName $newName}
 	if ($LASTEXITCODE) {
 		$2.Result = 'Ignore'
 		Show-FarMessage "Error on rename branch $oldName to $newName`n$res" -Caption FarGit -LeftAligned -IsWarning